Refactoring & trying to provide right stopwatch state for PreviewGameComponent
