+++ composeApp/src/commonMain/kotlin/com/arkivanov/minesweeper/game/stopwatch/StopwatchStateCalculator.kt
